+++ src/core/use-cases/list-todo/list-todo.ts
@@ -1,9 +1,14 @@
 import { Todo } from '../../entities';
 import { ITodoRepository } from '../../ports';
 import { IListTodoUseCase } from '..';
+import { UseCaseContext } from '../../../main/container';
 
 export class ListTodoUseCase implements IListTodoUseCase {
-  constructor(private todo_repository: ITodoRepository) {}
+  private todo_repository: ITodoRepository;
+
+  constructor(use_case_context: UseCaseContext) {
+    this.todo_repository = use_case_context.todo_repository;
+  }
 
   list(): Todo[] {
     return this.todo_repository.list();