+++ src/interface/http/controllers/list-todo/list-todo.ts
@@ -1,0 +1,30 @@
+import { Todo } from '../../../../core/entities';
+import { IListTodoUseCase } from '../../../../core/use-cases';
+
+import { HttpResponse, Controller, RouteConfig } from '../..';
+import { AuthMiddleware } from '../..';
+
+export class ListTodoController extends Controller {
+  route_configs: RouteConfig = {
+    method: 'get',
+    path: '/todos',
+    status_code: 200,
+    middlewares: [new AuthMiddleware()],
+  };
+
+  constructor(private list_todo_use_case: IListTodoUseCase) {
+    super();
+  }
+
+  async handle(): Promise<HttpResponse<Todo[]>> {
+    const todos = await this.list_todo_use_case.list();
+
+    return {
+      data: todos,
+    };
+  }
+
+  exception(error: Error): Error {
+    return error;
+  }
+}