+++ src/interface/http/middlewares/interface.ts
@@ -1,0 +1,5 @@
+import { HttpRequest, HttpResponse } from '..';
+
+export interface Middleware {
+  handle(req: HttpRequest, error?: Error): HttpResponse | Promise<void>;
+}