+++ src/main/modules/cli/cli.ts
@@ -1,5 +1,5 @@
 import readline from 'readline';
-import { ICreateTodoCommand } from '../../../interface';
+import { CreateTodoCommand, ICreateTodoCommand } from '../../../interface/cli';
 import { Container } from '../../container';
 import { Module } from '..';
 
@@ -17,7 +17,7 @@ export class CliModule implements Module {
         terminal: false,
       });
 
-    this.commands = container.create_todo_command;
+    this.commands = new CreateTodoCommand(container.create_todo_use_case);
   }
 
   start(): void {