+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Ongoing Changes
 
+- ref - container and cli module
+- add - module http
+- add - module cli
 - add - module and container layer
 - Add - interface layer
 - Add - infra layer